整合schema_manager到cache中
